Activation Layer

Layer factory function to create an activation layer. Activation functions can be used directly in CNTK, so there is no difference between `y = relu(x)` and `y = Activation(relu)(x)`. This layer is useful if one wants to configure the activation function with default <- options, or when its invocation should be named.

Activation(activation = activation_identity, name = "")

Arguments

activation
(defaults to `activation_identity`) – function to apply at the end, e.g. `relu`

name
(str, defaults to '') – the name of the function instance in the network

Value

A function that accepts one argument and applies the operation to it

Examples

```
model <- dense(500) %>% activation(op_relu)()
# is the same as
model <- dense(500) %>% op_relu
# and also the same as
model <- dense(500, activation=op_relu)
```

#> Error: attempt to use zero-length variable name