CNTKAxis

An axis object describes the axis of a variable and is used for specifying the axes parameters of certain functions such as reductions. Besides the static axes corresponding to each of the axes of the variable’s shape, variables of kind ‘input’ and any ‘output’ variables dependent on an ‘input’ variable also have two additional dynamic axes whose dimensions are known only when the variable is bound to actual data during compute time (viz. sequence axis and batch axis denoting the axis along which multiple sequences are batched).

CNTKAxis(...)

Details

Axis parameters can also be negative, which allows to refere axis starting from the last axis. Please be aware that Axis objects work in a column-major wise, as opposed to any other function in the library.

****** Properties: ******

is_ordered

is_static_axis

name

****** Associated Functions: ******

get_all_axes()

get_all_static_axes()

get_default_batch_axis()

create_new_leading_axis()

new_unique_dynamic_axis(name)

get_static_axis_index(axis, checked = TRUE)

unknown_dynamic_axes()