CNTK Function

Base class of all primitive tensor operators.

Function(...)

Arguments

...

Details

If it has only one output, one can invoke Variable methods on it, which it will relay to its only output.

Function objects can also be constructed directly from a Python lambda, by means of the @Function decorator. The Function‘s input signature is defined by the lambda.

****** Attributes: ******

arguments - List of all input variables of the Function that are not of type Parameter or Constant

attributes - List of the attributes of the function

block_arguments_mapping - The mapping from the arguments of the composite underlying this block function to the Variables that they are bound to in the outer graph of Functions that this block Function is part of.

block_root - The root of the Function graph underlying this block Function. Throws an exception if this is not a block Function.

constants - List of all Constant variables of this Function

inputs - List of variables that are inputs of this function. Note that ‘inputs’ here denotes all Variables that feed into this Function including any Parameter/Constant Variables that are children of this Function.

is_block - Boolean indicating if this Function is a block function which is basically a composite encapsulated as an opaque block which appears as a primitive during traversing the graph of Functions that this block is part of.

is_composite - Boolean indicating if this Function is a composite Function. A composite Function is a Function that is composed of primitive Functions.

is_primitive - Boolean indicating if this Function is a primitive Function. A primitive Function is the lowest level building block for composite Function graphs and is either a CNTK built-in operator, a composite Function encapsulated as a Block or a user-defined Function

name - Name of the Function

op_name - Name of the operation that this Function performs

output - The single output variable if there is only one, or raises an exception.

outputs - List consisting of all output variables of this Function.

parameters - List of all parameter variables of this Function.

placeholders - List of all placeholders variables of this Function.

root_function - The primitive function at the root of the graph of functions underlying this function.

signature - Signature of a Function. This is the $arguments list without placeholders that belong to an outer, not yet completed @Function def.

type - Get type of a Function's output

uid - The internally generated unique name of the Function

****** Associated Functions: ******

func_backward

func_clone

func_eval

func_find_all_with_name

func_find_by_name

func_forward

func_grad

func_load

register_udf_deserialize_callback

func_replace_placeholder

func_replace_placeholders

func_restore

func_save

func_set_attribute

func_test

func_train