Max-Pooling Layer Factory

Layer factory function to create a max-pooling layer. Like Convolution(), MaxPooling() processes items arranged on an N-dimensional grid, such as an image. Typically, each item is a vector. For each item, max-pooling computes the element-wise maximum over a window ("receptive field") of items surrounding the item’s position on the grid. The size (spatial extent) of the receptive field is given by filter_shape. E.g. for 2D pooling, filter_shape should be a tuple of two integers, such as (5,5).

MaxPooling(filter_shape, strides = 1, pad = FALSE, name = "")

Arguments

filter_shape
integer vector defining receptive field

strides
integer defining length of stride

pad
logical for whether or not the pooling operation should be shifted over the "valid" area of input

name
string for name of the function instance in the network