StreamDef

Configuration of a stream for use with the builtin Deserializers. The meanings of some configuration keys have a mild dependency on the exact deserializer, and certain keys are meaningless for certain deserializers.

StreamDef(field = NULL, shape = NULL, is_sparse = FALSE,
  transforms = NULL, context = NULL, scp = NULL, mlf = NULL,
  broadcast = NULL, defines_mb_size = FALSE)

Arguments

field
string defining the name of the stream

shape
integer defining the dimensions of the stream

is_sparse
logical for whether the data is sparse (FALSE by default)

transforms
list of transforms to be applied to the Deserializer

context
vector of length two defining whther reading in HTK data, (only supported by `HTKFeatureDeserializer`)

scp
list of `scp` files for HTK data

mlf
list `mlf` files for HTK data

broadcast
logical for whether the streams should be broadcast to the whole sequence

defines_mb_size
logical for whether this stream defines minibatch size

Value

A StreamDef object containing the stream dictionary

References

https://www.cntk.ai/pythondocs/cntk.io.html#cntk.io.StreamDef