Classification Error Between Target and Predicted

This operation computes the classification error. It finds the index of the highest value in the output_vector and compares it to the actual ground truth label (the index of the hot bit in the target vector). The result is a scalar (i.e., one by one matrix). This is often used as an evaluation criterion. It cannot be used as a training criterion though since the gradient is not defined for it.

classification_error(output_vector, target_vector, axis = -1, topN = 1,
  name = "")

Arguments

output_vector
the output values of the network

target_vector
one-hot encoding of target values

axis
integer for axis along which the classification error is computed

topN
integer

name
string (optional) the name of the Function instance in the network

References

https://www.cntk.ai/pythondocs/cntk.metrics.html#cntk.metrics.classification_error